tools/docs-code-compare/requirements.txt
PyYAML>=6.0.1


